+++ transform/lib/index.js
@@ -1,0 +1,5 @@
+import { Transform } from "assemblyscript/dist/transform.js";
+export default class BenchTransform extends Transform {
+    afterParse(_parser) {
+    }
+}